requirements.txt
fastmcp>=2.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
starlette>=0.27.0
httpx>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
tenacity>=8.0.0
PyJWT>=2.8.0